# feat: implement dynamic import caching for app store services